backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "travelbuddy"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "sqlalchemy[asyncio]>=2.0.30",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "pgvector>=0.2.5",
  "anthropic>=0.39",
  "langchain>=1.2",
  "langchain-anthropic>=1.4",
  "langchain-core>=1.3",
  "langgraph>=1.1",
  "httpx>=0.27",
  "python-jose[cryptography]>=3.3",
  "passlib[bcrypt]>=1.7",
  "bcrypt<4.1",
  "sentence-transformers>=2.7",
  "python-telegram-bot>=21.0",
  "python-multipart>=0.0.9",
  "sse-starlette>=2.1",
  "email-validator>=2.1",
  "numpy>=1.26"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5",
  "ruff>=0.4",
  "respx>=0.21"
]
whisper = [
  "faster-whisper>=1.0"
]

[tool.setuptools]
packages = { find = { where = ["."], include = ["app*"] } }

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E","F","I","B","UP","PL","RUF"]
ignore = ["PLR0913", "PLR2004", "PLR0912", "PLR0915"]
